feat(video): restructure video rendering code and show dimensions in sidebar
