+++ examples/basic/src/gatsby-theme-mdx-suite-core/components.js
@@ -1,7 +1,7 @@
 import defaultComponents from 'gatsby-theme-mdx-suite-base/src/gatsby-theme-mdx-suite-core/components'
 
 // Basics
-import Video from '@gatsby-mdx-suite/mdx-video/video-renderer'
+import Video from '@gatsby-mdx-suite/mdx-video/video'
 import YoutubeVideo from '@gatsby-mdx-suite/mdx-youtube/youtube-video'
 import VimeoVideo from '@gatsby-mdx-suite/mdx-vimeo/vimeo-video'
